decision-agent/agent-backend/agent-memory/pyproject.toml
[project]
name = "agent-memory"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "anthropic==0.18.1",
    "boto3==1.34.29",
    "dbutilsx==1.3.3",
    "fastapi==0.109.0",
    "groq==0.4.2",
    "httpx==0.25.2",
    "litellm==1.30.7",
    "numpy>=1.26",
    "ollama==0.1.6",
    "openai==1.14.2",
    "opensearch-py==2.8.0",
    "posthog==3.5.0",
    "pydantic>=2.9.0",
    "python-dotenv==1.0.1",
    "pytz==2024.1",
    "pyyaml==6.0.1",
    "PyMySQL==1.1.1",
    "dbutilsx==1.3.3",
    "qdrant-client==1.8.0",
    "redis==5.0.1",
    "requests==2.32.0",
    "setuptools==76.0.0",
    "sqlalchemy==1.4.23",
    "setuptools==76.0.0",
    "together==0.2.4",
    "uvicorn==0.27.0",
    "tomli>=2.0.0",
    "dbutilsx>=1.3.3",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.2",
    "pytest-asyncio==0.21.0",
    "pytest-cov==4.1.0",
    "mock==5.1.0",
    "pre-commit>=3.0.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=90",
]
asyncio_mode = "auto"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "slow: marks tests as slow running",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/env/*",
    "*/__pycache__/*",
    "*/mem0/*",
    "src/main.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
precision = 2


[tool.uv]
# 使用清华大学 PyPI 镜像源加速依赖下载
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "E501",  # line too long
    "E722",  # bare except
    "W605",  # invalid escape sequence (regex patterns)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "F401",  # unused import
    "F811",  # redefinition of unused
    "F821",  # undefined name
    "F841",  # unused variable
    "E721",  # type comparison
    "E731",  # lambda assignment
    "E741",  # ambiguous variable name
    "E402",  # import not at top
]
"src/main.py" = [
    "E402",  # import not at top (entry point script needs sys.path manipulation)
]
